Moves electron dependencies into desktop package
