+++ desktop/gulp/clean.js
@@ -1,0 +1,10 @@
+import gulp from 'gulp';
+import del from 'del';
+
+// Cleans out all build output folders folder
+export function all() {
+  return del([
+    'src',
+    'node_modules',
+  ]);
+}
