+++ desktop/gulp/shared.js
@@ -1,0 +1,2 @@
+import desktopPkgJson from './package.json';
+export { dektopPkgJson };
